refactor: Simplify navigation controls by delegating to CrossClickArea component
- Refactor NavigationControls.kt to act as a facade for the CrossClickArea component, enhancing code organization and maintainability.
- Remove redundant UI elements and logic, streamlining the navigation control implementation.
